+++ source/extensions/omni.isaac.ros2_bridge/python/nodes/OgnROS2RtxLidarHelper.py
@@ -24,13 +24,20 @@ class OgnROS2RtxLidarHelperInternalState(BaseWriterNode):
         self.viewport = None
         self.viewport_name = ""
         self.resetSimulationTimeOnStop = False
+        self.publishStepSize = 1
         super().__init__(initialize=False)
 
     def post_attach(self, writer, render_product):
         try:
+
+            omni.syntheticdata.SyntheticData.Get().set_node_attributes(
+                "PostProcessDispatch" + "IsaacSimulationGate", {"inputs:step": self.publishStepSize}, render_product
+            )
+
             omni.syntheticdata.SyntheticData.Get().set_node_attributes(
                 "IsaacReadSimulationTime", {"inputs:resetOnStop": self.resetSimulationTimeOnStop}, render_product
             )
+
         except:
             pass
 
@@ -81,6 +88,9 @@ class OgnROS2RtxLidarHelper:
                 db.per_instance_state.render_product_path = render_product_path
                 sensor_type = db.inputs.type
                 db.per_instance_state.resetSimulationTimeOnStop = db.inputs.resetSimulationTimeOnStop
+
+                db.per_instance_state.publishStepSize = db.inputs.frameSkipCount + 1
+
                 writer = None
 
                 time_type = ""